+++ primefaces/src/main/frontend/packages/components/src/columntoggler/columntoggler.widget.js
@@ -107,6 +107,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
             '<span class="ui-chkbox-icon ui-icon ui-icon-check"></span></div></div>' +
             '</li>');
         this.selectAllCheckbox.appendTo(this.itemContainer);
+        this.selectAllCheckbox = this.selectAllCheckbox.find('> .ui-chkbox > .ui-chkbox-box');
 
         //items
         for (var i = 0; i < this.columns.length; i++) {
@@ -144,7 +145,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
             }
 
             if (hidden) {
-                this.selectAllCheckbox.find('> .ui-chkbox > .ui-chkbox-box')
+                this.selectAllCheckbox
                     .removeClass('ui-state-active').attr('aria-label', this.getAriaLabel('unselectAll')).attr('aria-checked', 'false')
                     .children('.ui-chkbox-icon').addClass('ui-icon-blank').removeClass('ui-icon-check');
             }
@@ -192,13 +193,46 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
         });
 
         //select all
-        this.itemContainer.find('> .ui-columntoggler-all > .ui-chkbox > .ui-chkbox-box').on('mouseenter.columnToggler', function() {
+        this.selectAllCheckbox.on('mouseenter.columnToggler', function() {
             $(this).addClass('ui-state-hover');
         }).on('mouseleave.columnToggler', function() {
             $(this).removeClass('ui-state-hover');
+        }).on('focus.columnToggler', function() {
+            $(this).addClass('ui-state-focus');
+        }).on('blur.columnToggler', function(e) {
+            $(this).removeClass('ui-state-focus');
         }).on('click.columnToggler', function(e) {
             $this.toggleAll();
             e.preventDefault();
+        }).on('keydown.columnToggler', function(e) {
+            switch (e.code) {
+                case 'Tab':
+                    var inputs = $this.itemContainer.find('> .ui-columntoggler-item > .ui-chkbox > .ui-chkbox-box');
+                    var index = $(this).closest('li').index() - 1;
+                    var targetIndex = e.shiftKey ? index - 1 : index + 1;
+
+                    if (e.shiftKey) {
+                        if ($this.closer.is(':visible')) {
+                            $this.closer.trigger('focus');
+                        } else {
+                            inputs.eq($this.columns.length - 1).trigger('focus');
+                        }
+                    } else {
+                        inputs.eq(targetIndex).trigger('focus');
+                    }
+                    e.preventDefault();
+                    break;
+                case 'Enter':
+                case 'NumpadEnter':
+                case 'Space':
+                    $this.toggleAll();
+                    e.preventDefault();
+                    break;
+                case 'Escape':
+                    $this.closer.trigger('click');
+                    e.preventDefault();
+                    break;
+            }
         });
 
         //labels
@@ -236,7 +270,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
      */
     bindKeyEvents() {
         var $this = this,
-            inputs = this.itemContainer.find('> li > div.ui-chkbox > div.ui-chkbox-box');
+            inputs = this.itemContainer.find('> li:not(.ui-columntoggler-all) > div.ui-chkbox > div.ui-chkbox-box');
 
         this.trigger.on('focus.columnToggler', function() {
             $(this).addClass('ui-state-focus');
@@ -271,11 +305,12 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
         }).on('keydown.columnToggler', function(e) {
             switch (e.code) {
                 case 'Tab':
-                    var index = $(this).closest('li').index();
+                    // subtract 1 because the select all checkbox is the first item
+                    var index = $(this).closest('li').index() - 1;
                     var targetIndex = e.shiftKey ? index - 1 : index + 1;
 
                     if (e.shiftKey && index === 0 || !e.shiftKey && index === $this.columns.length - 1) {
-                        $this.closer.trigger('focus');
+                        $this.selectAllCheckbox.trigger('focus');
                     } else {
                         inputs.eq(targetIndex).trigger('focus');
                     }
@@ -285,6 +320,10 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
                 case 'NumpadEnter':
                 case 'Space':
                     $this.toggle($(this));
+                    e.preventDefault();
+                    break;
+                case 'Escape':
+                    $this.closer.trigger('click');
                     e.preventDefault();
                     break;
             }
@@ -303,6 +342,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
                 case 'Enter':
                 case 'NumpadEnter':
                 case 'Space':
+                case 'Escape':
                     $this.hide();
                     $this.trigger.trigger('focus');
                     e.preventDefault();
@@ -312,7 +352,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
                     if (e.shiftKey)
                         inputs.eq($this.columns.length - 1).trigger('focus');
                     else
-                        inputs.eq(0).trigger('focus');
+                        $this.selectAllCheckbox.trigger('focus');
 
                     e.preventDefault();
                     break;
@@ -338,7 +378,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
      * Toggles selecting or deselecting all columns.
      */
     toggleAll() {
-        if (this.selectAllCheckbox.find('> .ui-chkbox > .ui-chkbox-box').hasClass('ui-state-active')) {
+        if (this.selectAllCheckbox.hasClass('ui-state-active')) {
             this.uncheckAll();
         }
         else {
@@ -355,7 +395,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
             $this.check($(this));
         });
 
-        this.selectAllCheckbox.find('> .ui-chkbox > .ui-chkbox-box')
+        this.selectAllCheckbox
             .addClass('ui-state-active').attr('aria-label', this.getAriaLabel('selectAll')).attr('aria-checked', 'true')
             .children('.ui-chkbox-icon').addClass('ui-icon-check').removeClass('ui-icon-blank');
     }
@@ -436,7 +476,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
         }
         chkbox.removeClass('ui-state-active').children('.ui-chkbox-icon').addClass('ui-icon-blank').removeClass('ui-icon-check');
 
-        this.selectAllCheckbox.find('> .ui-chkbox > .ui-chkbox-box')
+        this.selectAllCheckbox
             .removeClass('ui-state-active').attr('aria-label', this.getAriaLabel('unselectAll')).attr('aria-checked', 'false')
             .children('.ui-chkbox-icon').addClass('ui-icon-blank').removeClass('ui-icon-check');
 
@@ -516,7 +556,12 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
         this.panel.show();
         this.visible = true;
         this.trigger.attr('aria-expanded', true);
-        this.closer.trigger('focus');
+        if (this.closer.is(':visible')) {
+            this.closer.trigger('focus');
+        }
+        else if (this.selectAllCheckbox.is(':visible')) {
+            this.selectAllCheckbox.trigger('focus');
+        }
     }
 
     /**
@@ -539,7 +584,7 @@ PrimeFaces.widget.ColumnToggler = class ColumnToggler extends PrimeFaces.widget.
      * @private
      */
     fireToggleEvent(visible, index) {
-        if (this.hasBehavior('toggle')) {
+        if (index >= 0 && this.hasBehavior('toggle')) {
             var ext = {
                 params: [
                     { name: this.id + '_visibility', value: visible ? 'VISIBLE' : 'HIDDEN' },