+++ primefaces/src/main/frontend/packages/components/src/datatable/datatable.widget.js
@@ -455,14 +455,14 @@ PrimeFaces.widget.DataTable = class DataTable extends PrimeFaces.widget.Deferred
      * Binds the change event listener and renders the paginator
      * @private
      */
-    bindPaginator() {
-        var _self = this;
+    bindPaginator:() {
+        var $this = this;
         this.cfg.paginator.paginate = function(newState) {
-            if(_self.cfg.clientCache) {
-                _self.loadDataWithCache(newState);
+            if($this.cfg.clientCache) {
+                $this.loadDataWithCache(newState);
             }
             else {
-                _self.paginate(newState);
+                $this.paginate(newState);
             }
         };
 
@@ -5764,6 +5764,15 @@ PrimeFaces.widget.DataTable = class DataTable extends PrimeFaces.widget.Deferred
         this.bodyTable.css('top', '0px');
         this.scrollBody.scrollTop(0);
         this.clearScrollState();
+    },
+
+    /**
+     * In case of a toggleable filter datatable toggle all filter input components
+     * @param {string | null} speed speed of fade animation. (see jquery fadeToggle method)
+     * @param {(() => void) | null} callback will be executed after animation is finished. (see jquery fadeToggle method)
+     */
+    toggleFilter: function(speed, callback) {
+        this.jq.find(".ui-column-filter, .ui-column-customfilter").fadeToggle(speed || 0, callback);
     }
 
 }